# backend/src/main/java/com/senior/spm/controller/request/CreateGroupRequest.java
package com.senior.spm.controller.request;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class CreateGroupRequest {

    @NotBlank(message = "Group name is required")
    private String groupName;
}