# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SPM API
  version: 1.0.0
  description: Senior Project Management system API
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
...
    GroupMember:
      type: object
      properties:
        studentId:
          type: string
        role:
          type: string
          enum: [TEAM_LEADER, MEMBER]
        joinedAt:
          type: string
          format: date-time
    CreateGroupRequest:
      type: object
      required: [groupName]
      properties:
        groupName:
          type: string
    CreateGroupResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupName:
          type: string
        termId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: '#/components/schemas/GroupMember'
    GroupDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupName:
          type: string
        termId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        jiraSpaceUrl:
          type: string
          nullable: true
        jiraProjectKey:
          type: string
          nullable: true
        jiraBound:
          type: boolean
        githubOrgName:
          type: string
          nullable: true
        githubBound:
          type: boolean
        advisorId:
          type: string
          format: uuid
          nullable: true
        members:
          type: array
          items:
            $ref: '#/components/schemas/GroupMember'
...
    SendInvitationRequest:
      type: object
      required: [targetStudentId]
      properties:
        targetStudentId:
          type: string
    InvitationResponse:
      type: object
      properties:
        invitationId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        targetStudentId:
          type: string
        status:
          type: string
          enum: [PENDING, ACCEPTED, DECLINED, AUTO_DENIED, CANCELLED]
        sentAt:
          type: string
          format: date-time
...
    BindJiraRequest:
      type: object
      required: [jiraSpaceUrl, jiraProjectKey, jiraApiToken]
      properties:
        jiraSpaceUrl:
          type: string
        jiraProjectKey:
          type: string
        jiraApiToken:
          type: string
    BindJiraResponse:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
        status:
          type: string
        jiraSpaceUrl:
          type: string
        jiraProjectKey:
          type: string
        jiraBound:
          type: boolean
    BindGithubRequest:
      type: object
      required: [githubOrgName, githubPat]
      properties:
        githubOrgName:
          type: string
        githubPat:
          type: string
    BindGithubResponse:
      type: object
      properties:
        groupId:
          type: string
          format: uuid
        status:
          type: string
        githubOrgName:
          type: string
        githubBound:
          type: boolean
...
security:
  - bearerAuth: []
paths:
...
  /groups:
    post:
      tags: [Student]
      summary: Validate Schedule & Create Group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGroupResponse'
        '400':
          description: Schedule window closed or student already in a group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '409':
          description: Group name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /groups/my:
    get:
      tags: [Student]
      summary: Get student's current group
      responses:
        '200':
          description: Returns the group details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDetailResponse'
        '404':
          description: Not a member of any group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
  /groups/{groupId}/invitations:
    post:
      tags: [Student]
      summary: Send Invitation
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendInvitationRequest'
      responses:
        '201':
          description: Invitation sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResponse'
        '400':
          description: Target student already in a group; group is DISBANDED; or (members + pending invitations) >= max team size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
          description: Caller is not the Team Leader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '409':
          description: Pending invitation already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    get:
      tags: [Student]
      summary: List group invitations
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Returns list of invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvitationResponse'
        '403':
          description: Caller is not the Team Leader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
  /groups/{groupId}/jira:
    post:
      tags: [Student]
      summary: Validate and Bind JIRA
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindJiraRequest'
      responses:
        '200':
          description: JIRA successfully bound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BindJiraResponse'
        '400':
          description: Group is DISBANDED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
          description: Caller is not the Team Leader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '422':
          description: JIRA Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /groups/{groupId}/github:
    post:
      tags: [Student]
      summary: Validate and Bind GitHub
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindGithubRequest'
      responses:
        '200':
          description: GitHub successfully bound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BindGithubResponse'
        '400':
          description: Group is DISBANDED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
          description: Caller is not the Team Leader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '422':
          description: GitHub Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...